* сохранение истории только в UTF-8 * исправлена прокрутка текста у групп * приборка

+++ trunk/src/conference/affiliation/AffiliationItem.java
@@ -25,7 +25,7 @@
  *
  */
 
-package conference.affiliation; 
+package conference.affiliation;
 
 import com.alsutton.jabber.JabberDataBlock;
 import images.RosterIcons;
@@ -41,8 +41,8 @@ public class AffiliationItem extends IconTextElement{
     public final static byte AFFILIATION_ADMIN=2;
     public final static byte AFFILIATION_MEMBER=3;
     public final static byte AFFILIATION_OUTCAST=4;
-    
-    public static String getAffiliationName(short index){
+
+    public static String getAffiliationName(int index){
         switch (index) {
             case AFFILIATION_OWNER: return "owner";
             case AFFILIATION_ADMIN: return "admin";
@@ -51,42 +51,42 @@ public class AffiliationItem extends IconTextElement{
         }
         return "none";
     };
-    
-    public int getImageIndex(){ 
+
+    public int getImageIndex(){
         switch (affiliation) {
             case AFFILIATION_OWNER: return RosterIcons.ICON_REGISTER_INDEX;
             case AFFILIATION_ADMIN: return RosterIcons.ICON_MODERATOR_INDEX;
             case AFFILIATION_MEMBER: return 0;
             case AFFILIATION_OUTCAST: return RosterIcons.ICON_ERROR_INDEX;
         }
-        return RosterIcons.ICON_INVISIBLE_INDEX; 
+        return RosterIcons.ICON_INVISIBLE_INDEX;
     }
 
     public String jid;
     public int affiliation;
     public String reason;
-        
+
     /** Creates a new instance of AffiliationItem */
     public AffiliationItem(String jid, String affiliation) {
         super(RosterIcons.getInstance());
         this.jid=jid;
         for (short index=1; index<5; index++) {
-            if (affiliation.equals(getAffiliationName(index))) 
+            if (affiliation.equals(getAffiliationName(index)))
                 this.affiliation=index;
         }
 	reason="";
     }
-    
+
     public AffiliationItem(JabberDataBlock item) {
         this(item.getAttribute("jid"), item.getAttribute("affiliation"));
         reason=item.getChildBlockText("reason");
-        
-	if (reason.length()==0) 
+
+	if (reason.length()==0)
             reason=null;
     }
-    
+
     public String toString() { return (reason!=null)? jid+" *" : jid; }
-    
+
     public int compare(IconTextElement temp2) {
         return jid.compareTo( ((AffiliationItem)temp2).jid);
     }
